CommercePreferences.BuyFromCheaperNeighbor -> LowestCost
Began work on changing the ReduceRecursively algorithm from stopping at
the first successful path it finds to fulfill the resource requirements
to continuing on until _every_ possible path has been checked and
picking the cheapest one possible (considering the cost of each
resource).  In this change I added the cost to the ResourceUsed.

The has/used traits have been changed to use a state system
(Unavailable, Unused, Used) but this is strictly a cosmetic change.